requirements.txt
numpy>=1.16.4
flax>=0.2.2
opencv-python>=4.4.0
pillow==9.5.0
matplotlib>=3.3.4
uvicorn==0.23.2
fastapi==0.94.0
python-dotenv==1.0.0
httpx==0.24.1